Move signing rest requests to RestClient
